feat(cli): add defineConfig for type helper
