+++ .gitignore
@@ -20,6 +20,7 @@ packages/varlet-cli/site/pc/*.routes.ts
 packages/varlet-cli/site/pc/routes.ts
 packages/varlet-cli/site/site.config.json
 packages/varlet-cli/site/*.site.config.json
+packages/varlet-cli/lib/bin.d.ts
 
 packages/varlet-ui/site
 packages/varlet-ui/lib
