+++ packages/varlet-cli/src/config/varlet.config.ts
@@ -3,7 +3,47 @@ import { merge } from 'lodash'
 import { VARLET_CONFIG, SITE_CONFIG } from '../shared/constant'
 import { outputFileSyncOnChange } from '../shared/fsUtils'
 
-export function getVarletConfig(emit = false): Record<string, any> {
+interface VarletConfig {
+  /**
+   * @default `Varlet`
+   * UI library name.
+   */
+  name?: string
+  /**
+   * @default `var`
+   * Component class name preffix
+   */
+  namesapce?: string
+  /**
+   * @default `localhost`
+   * Local dev server host
+   */
+  host?: string
+
+  /**
+   * @default `8080`
+   * Local dev server protcol
+   */
+  port?: number
+  logo?: string
+  defaultLanguage?: 'zh-CN' | 'en-US'
+  /**
+   * @default `false`
+   * Show mobile component on the right.
+   */
+  useMobile?: boolean
+  lightTheme?: Record<string, string>
+  darkTheme?: Record<string, string>
+  highlight?: { style: string }
+  analysis?: { baidu: string }
+  pc?: Record<string, any>
+  mobile?: Record<string, any>
+  moduleCompatible?: Record<string, string>
+}
+
+export const defineConfig = (conf: VarletConfig) => conf
+
+export function getVarletConfig(emit = false): any {
   let config: any = {}
 
   if (pathExistsSync(VARLET_CONFIG)) {
