+++ packages/varlet-cli/template/generators/config/default/base/varlet.config.js
@@ -1,4 +1,6 @@
-module.exports = {
+const { defineConfig } = require('@varlet/cli')
+
+module.exports = defineConfig({
   logo: './logo.svg',
   useMobile: true,
   pc: {
@@ -41,5 +43,5 @@ module.exports = {
       i18n: null,
       darkMode: null,
     },
-  },
-}
+  }
+})
