+++ packages/varlet-ui/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "module": "esnext",
+    "outDir": "./dist",
     "moduleResolution": "node",
     "strict": true,
     "downlevelIteration": true,
